Module google_commentanalyzer1_alpha1::api

Structs

AnalyzeCommentRequest
The comment analysis request message. LINT.IfChange

AnalyzeCommentResponse
The comment analysis response message.

ArticleAndParentComment
A type of context specific to a comment left on a single-threaded comment message board, where comments are either a top level comment or the child of a top level comment.

AttributeParameters
Configurable parameters for attribute scoring.

AttributeScores
This holds score values for a single attribute. It contains both per-span scores as well as an overall summary score..

CommentAnalyzeCall
Analyzes the provided text and returns scores for requested attributes.

CommentAnalyzer
Central instance to access all CommentAnalyzer related resource activities

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the CommentAnalyzer hub.

CommentSuggestscoreCall
Suggest comment scores as training data.

Context
Context is typically something that a Comment is referencing or replying to (such as an article, or previous comment). Note: Populate only ONE OF the following fields. The oneof syntax cannot be used because that would require nesting entries inside another message and breaking backwards compatibility. The server will return an error if more than one of the following fields is present.

Score
Analysis scores are described by a value and a ScoreType.

SpanScore
This is a single score for a given span of text.

SuggestCommentScoreRequest
The comment score suggestion request message.

SuggestCommentScoreResponse
The comment score suggestion response message.

TextEntry
Represents a body of text.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.